Moved searcher to util
