+++ ScorchApp/src/components/inventory/index.js
@@ -3,7 +3,6 @@ import WeaponInventory from './weapons/WeaponInventory'
 import AdventurerGearInventory from './adventurergear/AdventurerGearInventory'
 import QuiverInventory from './quiver/QuiverInventory'
 import ArmorInventory from './armor/ArmorInventory'
-import Searcher from './Searcher'
 import ItemCard from './ItemCard'
 
 export {
@@ -12,6 +11,5 @@ export {
     AdventurerGearInventory,
     QuiverInventory,
     ArmorInventory,
-    Searcher,
     ItemCard
 }
