+++ ScorchApp/src/components/util/index.js
@@ -1,5 +1,7 @@
 import Modal from './Modal.vue'
+import Searcher from './Searcher'
 
 export {
-    Modal
+    Modal,
+    Searcher    
 }
